+++ Sources/KeyboardKit/_Deprecated/View+Deprecated.swift
@@ -7,6 +7,8 @@ public extension View {
         for action: KeyboardAction,
         appearance: KeyboardAppearance,
         isPressed: Bool = false) -> some View {
-            systemKeyboardButtonStyle(for: action, appearance: appearance, isPressed: isPressed)
+            self.systemKeyboardButtonStyle(
+                style: appearance.systemKeyboardButtonStyle(for: action, isPressed: isPressed)
+            )
         }
 }